volume, nbd: connect UNIX socket to verified bind
Connect on AF_UNIX follows symlinks, and there is no way to switch this off.
This creates possibility for race condition, when symlink is replaced
after validation but before it resolve in connect.
Prevent this by binding socket via descriptor (move_mount) to protected directory
and connecting to this bind instead of original file. This guarantees
that kernel connects to the socket that we have validated.