cgroup: remove PortoTids
The initial problem that PortoTids tried to solve was that on kernels
prior commit \[1\] cgroup.procs/tasks could return 0. In this case porto
will call kill(0, SIGKILL) and in turn will kill itself.
Porto threads cannot occur in container cgroups, so there is no point
in checking that thread do not belong to portod process during
TCgroup::KillAll, checking that pid \> 0 is enough.

Also, usage of PortoTids is racy and in current implemetation has a
bug:
on thread exit its tid remain in PortoTid which can prevent us from
killing some other thread that recycled this tid.

\[1\] https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?h=2d4ecb030dcc90fb725ecbfc82ce5d6c37906e0e
65caa8b13eb3f77121aaa0e0b6f84fdf4895ec7d