cgroup: fix handling of empty cpuset in cgroup v2
In cgroup-v2, assigning an empty cpuset to a container currently
results in the cpuset being set to all present CPUs. This behavior can
lead to situation where we no longer can update container cpuset.

cgroup-v2 allows to set cgroup cpuset to value, that is not subset of
parent cgroup cpuset. However, prior to commit 1f1562fcd04a
(“cgroup/cpuset: Don’t let child cpusets restrict parent in default
hierarchy”), the kernel enforced restriction in opposite direction:
when updating a cpuset for cgroup A, all descendant cpusets had to
remain subsets of A cpuset. This restriction made it impossible to
change a cpuset for container with children.

To resolve this, this patch ensures that when a cpuset is left empty,
it inherits the parent’s cpuset, instead of setting it to all present
CPUs.
commit_hash:cf61558f547d2523f48be553d013cb8e220e93aa