volume: forbid shadowing of bind source
If a bind mount A is created and A's target is the source of another
bind B, a transitive bind A.source -> B.target is implicitly
created. If B's target points to place/storage/volume, the original
object may become inaccessible. To avoid losing access, shadowing of a
bind source (storage in terms of porto volumes) is disallowed.
commit_hash:07d8c3b6e9ee4c687d403175f788011bed19921f